+++ myuw/test/api/test_textbook.py
@@ -2,37 +2,54 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import json
-from myuw.test.api import require_url, MyuwApiTest
-from myuw.dao.registration import get_schedule_by_term
-from myuw.dao.instructor_schedule import get_instructor_schedule_by_term
-from myuw.views.api.textbook import (
-    get_textbook_by_schedule, _get_schedule_textbooks, index_by_sln)
+from myuw.test.api import MyuwApiTest, require_url, fdao_bookstore_override
 from myuw.test import get_request_with_user, get_request_with_date
 
 
-@require_url('myuw_current_book')
-class TestInstructorTextbookCur(MyuwApiTest):
+VERBACOMPARE_URL_PREFIX = 'http://uw-seattle.verbacompare.com'
+IMAGE_URL_PREFIX = 'www7.bookstore.washington.edu/MyUWImage.taf'
 
-    def test_get(self):
-        self.set_user('billpce')
-        self.set_date("2013-10-16 09:00:00")
-        resp = self.get_response_by_reverse('myuw_current_book')
-        data = json.loads(resp.content)
-        self.assertEqual(data, {'21838': []})
 
-    def test_get_uwt_inst_textbook(self):
-        # MUWM-5311: uwt no longer has books
-        req = get_request_with_user('billtac',
-                                    get_request_with_date("2013-04-01"))
-        schedule = get_instructor_schedule_by_term(req)
-        bkData = get_textbook_by_schedule(schedule)
-        self.assertEqual(bkData, {})
-        self.assertEqual(index_by_sln(bkData), {})
+@fdao_bookstore_override
+class TestApiBooks(MyuwApiTest):
+    '''Tests textbooks api'''
 
-        data = _get_schedule_textbooks(schedule)
-        self.assertEqual(data, {})
+    @require_url('myuw_home')
+    def test_javerage_books(self):
+        self.set_user('javerage')
+        response = self.get_response_by_reverse(
+            'myuw_book_api',
+            kwargs={'year': 2013,
+                    'quarter': 'spring',
+                    'summer_term': ''})
+        self.assertEquals(response.status_code, 200)
 
-        self.set_user('billtac')
-        resp = self.get_response_by_reverse('myuw_current_book')
-        data = json.loads(resp.content)
-        self.assertEqual(data, {})
+        data = json.loads(response.content)
+
+        self.assertEquals(data["order_url"],
+                          ('http://www.ubookstore.com/adoption-search' +
+                           '-results?ccid=9335,1132,5320,2230,4405'))
+
+        self.assertGreaterEqual(len(data['18532']), 1)
+        book = data['18532'][0]
+        self.assertEquals(
+            book["cover_image_url"],
+            ("%s?isbn=9780878935970&key=46c9ef715edb2ec69517e2c8e6ec9c18" %
+             IMAGE_URL_PREFIX))
+        self.assertEquals(len(book["authors"]), 1)
+        self.assertTrue(book["is_required"])
+        self.assertIsNone(book["price"])
+        self.assertIsNone(book["used_price"])
+        self.assertEquals(book["isbn"], '9780878935970')
+        self.assertEquals(book["notes"], 'required')
+        self.assertIsNone(book["price"])
+
+    @require_url('myuw_home')
+    def test_no_sche_books(self):
+        self.set_user('billtac')  # MUWM-5311
+        response = self.get_response_by_reverse(
+            'myuw_book_api',
+            kwargs={'year': 2013,
+                    'quarter': 'spring',
+                    'summer_term': ''})
+        self.assertEquals(response.status_code, 200)
